pyproject.toml
[project]
name = "karp-lex"
version = "0.1.0"
description = "The domain of the lexical platform Karp"
authors = [
    { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
]
dependencies = ["pydantic>=1.10.2", "ulid-py>=1.1.0"]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
[project.optional-dependencies]


[project.urls]
homepage = "https://github.com/spraakbanken/karp-lex"
# documentation = "https://readthedocs.org"
repository = "https://github.com/spraakbanken/karp-lex"
changelog = "https://github.com/spraakbanken/karp-lex/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/spraakbanken/karp-lex/issues"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.1.3",
    "pytest-cov>=4.0.0",
    "black>=22.8.0",
    "bump2version>=1.0.1",
    "pylint>=2.15.3",
    "mypy>=0.982",
]
doc = ["mkdocs"]
